+++ RELEASE-NOTES.md
@@ -36,6 +36,7 @@
 - **Form Validation** - Revalidating a field `on: blur` could cause fields not yet interacted with to be validated #3606
 - **Rail** - Fixed incorrect width for `close rail` and `very close rail` caused by variable addition with mixed units `px` + `em` #3835
 - **Search** - A previous unfinished XHR query aborting would cause the next query to fail #2779
+- **Search** - Fixed an issue where `onResult` returning `false` would not prevent the search menu from hiding. Clicking on an empty results message will also no longer close the search results. #3856 #3870
 - **Video** - Fixed issue with `change` behavior not working properly when correctly to change videos.
 
 **Bugs**